Split up views.py.
